set start to 0 on facet selection
Changed the signature to search(start) which will set this.searchStart
if the value is not undefined. A 0 is passed when a facet select is
made.

This addressed the issue without causing a regression with paging,
though I feel like the issue is in the paging calculation in the api
search promise success case.

refs #39