Cleanup of refactor adding sample plainhtml plugin
Reworked the repo - shifted the editor to the webeditor directory, left the plugin directory as is, but added the sample applications using the belmgr plugins.  Refactored plugin and webeditor to allow for easy local development and direct use of the belmgr plugin as a standard Aurelia plugin in the webeditor code as well as the sample-aureliaplugin application.

For local dev however, we build the plugin and use sym links to make the plugin code available in the jspm_packages npm directory for use in webeditor and the sample applications.